removed namespace phantom.

+++ game/src/components/Console.cpp
@@ -1,5 +1,7 @@
 #include "Console.h"
 
+#include <gl/freeglut.h>
+
 Console* Console::INSTANCE = 0;
 
 void Console::log(string log) {
@@ -45,7 +47,7 @@ void Console::renderText(int offset, Color color) {
     float lineOffset = lineheight;
 
     for(string& log : _logs) {
-        g.text(offset, lineOffset + offset, GLUT_BITMAP_HELVETICA_10, &log);
+        g.text(static_cast<float>(offset), lineOffset + static_cast<float>(offset), GLUT_BITMAP_HELVETICA_10, &log);
         lineOffset += lineheight;
     }
